requirements.txt
streamlit==1.31.0
opencv-python-headless==4.9.0.80
numpy==1.26.0
pandas==2.1.0
python-dotenv==1.0.0
Pillow==10.0.0
easyocr==1.7.1
pytesseract==0.3.10
spacy==3.7.2
transformers==4.35.2
torch==2.1.0
torchvision==0.16.0
python-multipart==0.0.6
python-docx==0.8.11
openpyxl==3.1.2
python-dateutil==2.8.2
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'
